feat(core): CATALYST-201 fetch brand/brands with new client
