+++ packages/client/src/client.ts
@@ -15,8 +15,6 @@ import { login } from './mutations/login';
 import { updateCartLineItem } from './mutations/updateCartLineItem';
 // Queries
 import { getBestSellingProducts } from './queries/getBestSellingProducts';
-import { getBrand } from './queries/getBrand';
-import { getBrands } from './queries/getBrands';
 import { getCart } from './queries/getCart';
 import { getCategory } from './queries/getCategory';
 import { getCategoryTree } from './queries/getCategoryTree';
@@ -46,14 +44,6 @@ class Client<CustomRequestInit extends FetcherRequestInit = FetcherRequestInit> 
 
   getBestSellingProducts(...args: PublicParams<typeof getBestSellingProducts<CustomRequestInit>>) {
     return getBestSellingProducts(this.fetch, ...args);
-  }
-
-  getBrand(...args: PublicParams<typeof getBrand<CustomRequestInit>>) {
-    return getBrand(this.fetch, ...args);
-  }
-
-  getBrands(...args: PublicParams<typeof getBrands<CustomRequestInit>>) {
-    return getBrands(this.fetch, ...args);
   }
 
   getCategory(...args: PublicParams<typeof getCategory<CustomRequestInit>>) {
